Hopefuly fixed convert

+++ src/NadekoBot/_Extensions/Extensions.cs
@@ -23,7 +23,7 @@ namespace NadekoBot.Extensions
             http.DefaultRequestHeaders.Add("Accept", "text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8");
         }
 
-        public static bool IsInteger(this double number) => number == Math.Truncate(number);
+        public static bool IsInteger(this decimal number) => number == Math.Truncate(number);
 
         public static string SanitizeMentions(this string str) => 
             str.Replace("@everyone", "@everyοne").Replace("@here", "@һere");
